refactor: tmpl package
